swupdate classes: fix hash mismatch on master
The fix discussed in [1] resolves build issues on styhead but
introduces an issue which causes hash mismatches.

The first build succeeds to generate a working .swu, however
future builds fail with hash mismatch and a message like this:

```
[ERROR] : SWUPDATE failed [0] ERROR cpio_utils.c : __swupdate_copy : 670 : HASH mismatch : 5bb39965bfcedd87fc02bd51fac6636b924aed606d0984dcebae6a3250b01af6 <--> ff70a8dbc50a9ce2bad724554f5f14b8c933ad87fc544d054142755e8ef07f1e
[ERROR] : SWUPDATE failed [0] ERROR archive_handler.c : install_archive_image : 344 : Error copying extracted file
```

This happens because the modifications made to files in the
S directory only happen once.  This means there's a single
conversion of `$swupdate_get_sha256(` to the file sha, and future
builds use the hash previously placed in the file in the S directory
which is no longer correct when the rootfs file hash changes.

It would probably be possible to make a simpler change here which
splits the UNPACKDIR and S and continues to make runtime
edits in S.  However, it seems the intent of S and/or UNPACKDIR
is to contain unmodified source as discussed in [1].

This change instead:
1. Moves the sw-description file to WORKDIR for editing.
2. Continues to use S as source location for sw-description
and other source files which are not edited.
3. Makes all edits in WORKDIR.
4. When preparing the cpio, prepares from WORKDIR instead of
S. using any existing files in WORKDIR if they exist.  If they
do not exist, symlinks the relevant files in S instead.
6. Adds the -L option to cpio to prevent adding symlinks to the
cpio file, and avoid the need to copy from S to WORKDIR when
preparing the cpio.

1: https://groups.google.com/g/swupdate/c/8K-9H7C9o5E/m/a9fqhIOjAAAJ

Signed-off-by: Dan Walkes <danwalkes@trellis-logic.com>